backend/requirements.txt
Flask
sentence-transformers
numpy
requests
python-dotenv